backend/requirements-lock.txt
# VisionAutoTest Backend production lock file
# Exact runtime dependency set for deployment/runtime environments.
# After installing, run: `playwright install chromium`

albucore==0.0.24
albumentations==2.0.8
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==45.0.7
Cython==3.2.4
fastapi==0.135.2
fire==0.7.1
fonttools==4.62.1
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
ImageIO==2.37.3
lazy-loader==0.5
lmdb==2.1.1
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
networkx==3.6.1
numpy==2.4.3
opencv-contrib-python==4.13.0.92
opencv-python-headless==4.13.0.92
opencv-python==4.13.0.92
opt-einsum==3.3.0
paddleocr==2.10.0
paddlepaddle==3.3.1
pillow==12.1.1
playwright==1.58.0
protobuf==7.34.1
psycopg-binary==3.3.3
psycopg==3.3.3
pycparser==3.0
pyclipper==1.4.0
pydantic-settings==2.13.1
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
PyJWT==2.12.1
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.33.0
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.17.1
shapely==2.1.2
simsimd==6.5.16
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==1.0.0
stringzilla==4.6.0
termcolor==3.3.0
tifffile==2026.3.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
